Adds authentication method and minor comment edition.
